Fixed bug on android-gradle-plugin:0.6.+
Fixed bug causing this plugin not to work with com.android.tools.build:gradle:0.6.+

Build was failing with following message
> Failed to notify project evaluation listener.
> Task with name 'compileDebugJava' not found in project ':Quick Timer'.

The reason of error is that gradle plugin 0.6 generates tasks named compileDebug, compileRelease and compileTest, but not compileDebugJava, compileReleaseJava and compileTestJava.
My solution is generic and will work on any plugin version. I'm filtering tasks by task type -- JavaCompile and then filtering it by beginning of it's name.