Major refactoring of android plugin.
The method for modifying the javaCompile task is now *way* less hackey. In fact,
it is much closer to the original way that it was done. I had originally
abandoned this approach because it was breaking increment compilation is some
weird ways. However, I think I have solved it.

What this means for you: It is now less fickle of plugin application order and
way less likely to break other plugins.

Note: I also threw in the fix for #101