This quick patch fixes #134
When code generation tasks, which change `sourceSet` during the execution are involved we cannot statically set `enabled` flag during retrolambda task creation. As a quick patch I suggest to leave it always enabled. In the body of the task itself it can check whether the sourceSet is empty and thus skip an execution